frontend/requirements.txt
pandas
plotly-express
datetime
python-dotenv
pymongo
beautifulsoup4
logging
requests